Move process class types into fdbclient
