+++ fdbrpc/include/fdbrpc/Locality.h
@@ -24,135 +24,6 @@
 
 #include "flow/flow.h"
 
-struct ProcessClass {
-	constexpr static FileIdentifier file_identifier = 6697257;
-	// This enum is stored in restartInfo.ini for upgrade tests, so be very careful about changing the existing items!
-	enum ClassType {
-		UnsetClass,
-		StorageClass,
-		TransactionClass,
-		ResolutionClass,
-		TesterClass,
-		CommitProxyClass,
-		MasterClass,
-		StatelessClass,
-		LogClass,
-		ClusterControllerClass,
-		LogRouterClass,
-		FastRestoreClass, // deprecated
-		DataDistributorClass,
-		CoordinatorClass,
-		RatekeeperClass,
-		RemovedClassPlaceholder1, // removing the name of removed functionality, but don't renumber subsequent entries.
-		BackupClass,
-		GrvProxyClass,
-		BlobManagerClass,
-		BlobWorkerClass,
-		EncryptKeyProxyClass,
-		ConsistencyScanClass,
-		BlobMigratorClass,
-		SimHTTPServerClass,
-		InvalidClass = -1
-	};
-
-	// class is serialized by enum value, so it's important not to change the
-	// enum value of a class. New classes should only be added to the end.
-	static_assert(ProcessClass::UnsetClass == 0);
-	static_assert(ProcessClass::StorageClass == 1);
-	static_assert(ProcessClass::TransactionClass == 2);
-	static_assert(ProcessClass::ResolutionClass == 3);
-	static_assert(ProcessClass::TesterClass == 4);
-	static_assert(ProcessClass::CommitProxyClass == 5);
-	static_assert(ProcessClass::MasterClass == 6);
-	static_assert(ProcessClass::StatelessClass == 7);
-	static_assert(ProcessClass::LogClass == 8);
-	static_assert(ProcessClass::ClusterControllerClass == 9);
-	static_assert(ProcessClass::LogRouterClass == 10);
-	static_assert(ProcessClass::FastRestoreClass == 11);
-	static_assert(ProcessClass::DataDistributorClass == 12);
-	static_assert(ProcessClass::CoordinatorClass == 13);
-	static_assert(ProcessClass::RatekeeperClass == 14);
-	// intentional gap at 15
-	static_assert(ProcessClass::BackupClass == 16);
-	static_assert(ProcessClass::GrvProxyClass == 17);
-	static_assert(ProcessClass::BlobManagerClass == 18);
-	static_assert(ProcessClass::BlobWorkerClass == 19);
-	static_assert(ProcessClass::EncryptKeyProxyClass == 20);
-	static_assert(ProcessClass::ConsistencyScanClass == 21);
-	static_assert(ProcessClass::BlobMigratorClass == 22);
-	static_assert(ProcessClass::SimHTTPServerClass == 23);
-	static_assert(ProcessClass::InvalidClass == -1);
-
-	enum Fitness {
-		BestFit,
-		GoodFit,
-		UnsetFit,
-		OkayFit,
-		WorstFit,
-		ExcludeFit,
-		NeverAssign
-	}; // cannot be larger than 7 because of leader election mask
-	enum ClusterRole {
-		Storage,
-		TLog,
-		CommitProxy,
-		GrvProxy,
-		Master,
-		Resolver,
-		LogRouter,
-		ClusterController,
-		DataDistributor,
-		Ratekeeper,
-		ConsistencyScan,
-		BlobManager,
-		BlobWorker,
-		BlobMigrator,
-		RemovedRolePlaceholder1,
-		Backup,
-		EncryptKeyProxy,
-		Worker, // used for actor lineage tracking
-		NoRole
-	};
-	enum ClassSource { CommandLineSource, AutoSource, DBSource, InvalidSource = -1 };
-	int16_t _class;
-	int16_t _source;
-
-	// source is serialized by enum value, so it's important not to change the
-	// enum value of a source. New sources should only be added to the end.
-	static_assert(ProcessClass::CommandLineSource == 0);
-	static_assert(ProcessClass::AutoSource == 1);
-	static_assert(ProcessClass::DBSource == 2);
-
-public:
-	ProcessClass() : _class(UnsetClass), _source(CommandLineSource) {}
-	ProcessClass(ClassType type, ClassSource source) : _class(type), _source(source) {}
-	explicit ProcessClass(std::string s, ClassSource source);
-
-	ProcessClass(std::string classStr, std::string sourceStr);
-
-	ClassType classType() const { return (ClassType)_class; }
-	ClassSource classSource() const { return (ClassSource)_source; }
-
-	bool operator==(const ClassType& rhs) const { return _class == rhs; }
-	bool operator!=(const ClassType& rhs) const { return _class != rhs; }
-
-	bool operator==(const ProcessClass& rhs) const { return _class == rhs._class && _source == rhs._source; }
-	bool operator!=(const ProcessClass& rhs) const { return _class != rhs._class || _source != rhs._source; }
-
-	std::string toString() const;
-
-	std::string sourceString() const;
-
-	Fitness machineClassFitness(ClusterRole role) const;
-
-	// To change this serialization, ProtocolVersion::ProcessClassValue must be updated, and downgrades need to be
-	// considered
-	template <class Ar>
-	void serialize(Ar& ar) {
-		serializer(ar, _class, _source);
-	}
-};
-
 struct LocalityData {
 	std::map<Standalone<StringRef>, Optional<Standalone<StringRef>>> _data;
 
@@ -300,39 +171,6 @@ inline std::string describeZones(std::vector<LocalityData> const& items, int max
 inline std::string describeDataHalls(std::vector<LocalityData> const& items, int max_items = -1) {
 	return describe(items, LocalityData::keyDataHallId, max_items);
 }
-
-struct ProcessData {
-	LocalityData locality;
-	ProcessClass processClass;
-	NetworkAddress address;
-	Optional<NetworkAddress> grpcAddress;
-
-	ProcessData() {}
-	ProcessData(LocalityData locality,
-	            ProcessClass processClass,
-	            NetworkAddress address,
-	            Optional<NetworkAddress> grpcAddress)
-	  : locality(locality), processClass(processClass), address(address), grpcAddress(grpcAddress) {}
-
-	// To change this serialization, ProtocolVersion::WorkerListValue must be updated, and downgrades need to be
-	// considered
-	template <class Ar>
-	void serialize(Ar& ar) {
-		serializer(ar, locality, processClass, address);
-
-		if constexpr (!is_fb_function<Ar>) {
-			if (ar.protocolVersion().hasGrpcEndpoint()) {
-				serializer(ar, grpcAddress);
-			}
-		} else {
-			serializer(ar, grpcAddress);
-		}
-	}
-
-	struct sort_by_address {
-		bool operator()(ProcessData const& a, ProcessData const& b) const { return a.address < b.address; }
-	};
-};
 
 template <class Interface, class Enable = void>
 struct LBLocalityData {
