+++ fdbrpc/include/fdbrpc/SimulatorProcessMetadata.h
@@ -1,0 +1,44 @@
+/*
+ * SimulatorProcessMetadata.h
+ *
+ * This source file is part of the FoundationDB open source project
+ *
+ * Copyright 2013-2026 Apple Inc. and the FoundationDB project authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef FDBRPC_SIMULATOR_PROCESS_METADATA_H
+#define FDBRPC_SIMULATOR_PROCESS_METADATA_H
+#pragma once
+
+#include <string>
+#include <utility>
+
+#include "flow/FastRef.h"
+
+namespace simulator {
+
+struct ProcessInfoMetadata : ReferenceCounted<ProcessInfoMetadata> {
+	std::string role;
+	bool excludeFromMachineKill;
+
+	explicit ProcessInfoMetadata(std::string role, bool excludeFromMachineKill = false)
+	  : role(std::move(role)), excludeFromMachineKill(excludeFromMachineKill) {}
+
+	virtual ~ProcessInfoMetadata() = default;
+};
+
+} // namespace simulator
+
+#endif // FDBRPC_SIMULATOR_PROCESS_METADATA_H
