+++ fdbserver/workloads/RemoveServersSafely.cpp
@@ -19,6 +19,7 @@
  */
 
 #include "fdbclient/NativeAPI.actor.h"
+#include "fdbclient/FDBSimulatorProcessInfo.h"
 #include "fdbserver/core/TesterInterface.h"
 #include "fdbserver/core/WorkerInterface.actor.h"
 #include "fdbserver/core/FDBSimulationPolicy.h"
@@ -776,7 +777,7 @@ struct RemoveServersSafelyWorkload : TestWorkload {
 		std::vector<ISimulator::ProcessInfo*> machines;
 		std::vector<ISimulator::ProcessInfo*> all = g_simulator->getAllProcesses();
 		for (int i = 0; i < all.size(); i++) {
-			if (all[i]->name == std::string("Server") && all[i]->isAvailableClass()) {
+			if (all[i]->name == std::string("Server") && isAvailableSimulatorProcessClass(all[i])) {
 				machines.push_back(all[i]);
 			}
 		}
